Telephony: Convert subscription ID to int from long

+++ telecomm/java/android/telecom/TelecomManager.java
@@ -930,10 +930,10 @@ public class TelecomManager {
      * Returns current active subscription.
      * @hide
      */
-    public long getActiveSubscription() {
+    public int getActiveSubscription() {
         try {
             if (isServiceConnected()) {
-                return getTelecomService().getActiveSubscription();
+                return (int)getTelecomService().getActiveSubscription();
             }
         } catch (RemoteException e) {
             Log.e(TAG, "RemoteException attempting to get the active subsription.", e);
@@ -945,7 +945,7 @@ public class TelecomManager {
      * switches to other active subscription.
      * @hide
      */
-    public void switchToOtherActiveSub(long subId) {
+    public void switchToOtherActiveSub(int subId) {
         try {
             if (isServiceConnected()) {
                 getTelecomService().switchToOtherActiveSub(subId);
